+++ src/OpenTelemetry.Extensions/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Update BaggageActivityProcessor to require baggage key predicate.
+  ([#1816](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1816))
+
 * Updated OpenTelemetry core component version(s) to `1.9.0`.
   ([#1888](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1888))